+++ Libplanet/Blockchain/BlockChain.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Numerics;
@@ -676,6 +677,7 @@ namespace Libplanet.Blockchain
             }
         }
 
+#pragma warning disable MEN003
         /// <summary>
         /// Mines a next <see cref="Block{T}"/> using staged <see cref="Transaction{T}"/>s,
         /// and then <see cref="Append(Block{T}, StateCompleterSet{T}?)"/> it to the chain
@@ -719,16 +721,29 @@ namespace Libplanet.Blockchain
             long difficulty = Policy.GetNextBlockDifficulty(this);
             HashDigest<SHA256>? prevHash = Store.IndexBlockHash(Id, index - 1);
 
-            Transaction<T>[] stagedTransactions = Store
-                .IterateStagedTransactionIds()
-                .Select(Store.GetTransaction<T>)
-                .GroupBy(tx => tx.Signer)
-                .SelectMany(grp => grp.Select((tx, idx) => (tx, idx)))
-                .OrderBy(t => t.Item2)
-                .Select(t => t.Item1)
-                .ToArray();
+            int sessionId = new System.Random().Next();
+            int procId = Process.GetCurrentProcess().Id;
+            _logger.Debug(
+                "Start to mine a block #{Index} [difficulty: {Difficulty}; prev: {prevHash}; " +
+                "session: {SessionId}; proc: {ProcessId}]... ",
+                index,
+                difficulty,
+                prevHash,
+                sessionId,
+                procId
+            );
+
+            ImmutableArray<Transaction<T>> stagedTransactions = ListStagedTransactions();
+            _logger.Debug(
+                "There are {Transactions} staged transactions.",
+                stagedTransactions.Length
+            );
 
             var transactionsToMine = new List<Transaction<T>>();
+            int i = 0;
+
+            // FIXME: The tx collection timeout should be configurable.
+            DateTimeOffset timeout = DateTimeOffset.UtcNow + TimeSpan.FromSeconds(4);
 
             // Makes an empty block to estimate the length of bytes without transactions.
             var estimatedBytes = new Block<T>(
@@ -746,27 +761,44 @@ namespace Libplanet.Blockchain
 
             foreach (Transaction<T> tx in stagedTransactions)
             {
+                _logger.Verbose(
+                    "Preparing mining a block #{Index}; validating a tx {Index}/{Total} " +
+                    "{Transaction}...",
+                    index,
+                    ++i,
+                    stagedTransactions.Length,
+                    tx.Id
+                );
+
+                if (transactionsToMine.Count >= maxTransactions)
+                {
+                    _logger.Information(
+                        "Not all staged transactions will be included in a block #{Index} to " +
+                        "be mined by {Miner}, because it reaches the maximum number of " +
+                        "acceptable transactions: {MaxTransactions}",
+                        index,
+                        miner,
+                        maxTransactions
+                    );
+                    break;
+                }
+
                 if (!Policy.DoesTransactionFollowsPolicy(tx, this))
                 {
+                    _logger.Debug(
+                        "Unstage the tx {Index}/{Total} {Transaction} as it doesn't follow policy.",
+                        i,
+                        stagedTransactions.Length,
+                        tx.Id
+                    );
                     UnstageTransaction(tx);
+                    continue;
                 }
-                else if (!skippedSigners.Contains(tx.Signer) &&
-                         Store.GetTxNonce(Id, tx.Signer) <= tx.Nonce &&
-                         tx.Nonce < GetNextTxNonce(tx.Signer))
-                {
-                    if (transactionsToMine.Count >= maxTransactions)
-                    {
-                        _logger.Information(
-                            "Not all staged transactions will be included in a block #{Index} to " +
-                            "be mined by {Miner}, because it reaches the maximum number of " +
-                            "acceptable transactions: {MaxTransactions}",
-                            index,
-                            miner,
-                            maxTransactions
-                        );
-                        break;
-                    }
 
+                long storeNonce = Store.GetTxNonce(Id, tx.Signer);
+                long nextNonce = GetNextTxNonce(tx.Signer);
+                if (storeNonce <= tx.Nonce && tx.Nonce < nextNonce)
+                {
                     if (estimatedBytes + tx.BytesLength > maxBlockBytes)
                     {
                         // Once someone's tx is excluded from a block, their later txs are also all
@@ -786,6 +818,27 @@ namespace Libplanet.Blockchain
 
                     transactionsToMine.Add(tx);
                     estimatedBytes += tx.BytesLength;
+                }
+                else
+                {
+                    _logger.Debug(
+                        "Tx {Index}/{Total} {Transaction} has an invalid nonce: {Nonce} " +
+                        "({Signer}).",
+                        i,
+                        stagedTransactions.Length,
+                        tx.Id,
+                        tx.Nonce,
+                        tx.Signer
+                    );
+                }
+
+                if (timeout < DateTimeOffset.UtcNow)
+                {
+                    _logger.Debug(
+                        "Reached the time limit to collect staged transactions; other staged " +
+                        "transactions will be mined later."
+                    );
+                    break;
                 }
             }
 
@@ -847,13 +900,34 @@ namespace Libplanet.Blockchain
                 block = new Block<T>(block, trieStateStore.GetRootHash(block.Hash));
             }
 
+            _logger.Debug(
+                "Mined a block #{Index} [difficulty: {Difficulty}; prev: {prevHash}; " +
+                "session: {SessionId}; proc: {ProcessId}].",
+                index,
+                difficulty,
+                prevHash,
+                sessionId,
+                procId
+            );
+
             if (append)
             {
                 Append(block, currentTime);
+
+                _logger.Debug(
+                    "Appended a block #{Index} [difficulty: {Difficulty}; prev: {prevHash}; " +
+                    "session: {SessionId}; proc: {ProcessId}].",
+                    index,
+                    difficulty,
+                    prevHash,
+                    sessionId,
+                    procId
+                );
             }
 
             return block;
         }
+#pragma warning restore MEN003
 
         /// <summary>
         /// Mines a next <see cref="Block{T}"/> using staged <see cref="Transaction{T}"/>s,
@@ -1626,6 +1700,27 @@ namespace Libplanet.Blockchain
             }
         }
 #pragma warning restore MEN003
+
+        internal ImmutableArray<Transaction<T>> ListStagedTransactions()
+        {
+            Transaction<T>[] txs = Store
+                .IterateStagedTransactionIds()
+                .Select(Store.GetTransaction<T>)
+                .ToArray();
+
+            Dictionary<Address, LinkedList<Transaction<T>>> seats = txs
+                .GroupBy(tx => tx.Signer)
+                .Select(g => (g.Key, new LinkedList<Transaction<T>>(g.OrderBy(tx => tx.Nonce))))
+                .ToDictionary(pair => pair.Item1, pair => pair.Item2);
+
+            return txs.Select(tx =>
+            {
+                LinkedList<Transaction<T>> seat = seats[tx.Signer];
+                Transaction<T> first = seat.First.Value;
+                seat.RemoveFirst();
+                return first;
+            }).ToImmutableArray();
+        }
 
         internal IImmutableSet<TxId> GetStagedTransactionIds()
         {